Renamed Planner to weather app only
