+++ src/App.tsx
@@ -21,8 +21,8 @@ import Register from "./pages/auth/Register";
 import Profile from "./pages/auth/Profile";
 import UpdateRessource from "./pages/dashboard/UpdateRessource";
 import RessourcesList from "./pages/dashboard/RessourcesList";
-import Planner from "./pages/planner/Planner";
-import PlannerApp from "./pages/planner/PlannerApp";
+import PlannerApp from "./pages/planner/WeatherApp";
+import WeatherApp from "./pages/planner/WeatherApp";
 
 function App() {
   return (
@@ -41,7 +41,7 @@ function App() {
           <Route path={routes.hub.path} element={<Layout component={<Hub />} />} />
           {/* <Route path={routes.planner.path} element={<Layout component={<Planner />} />} /> */}
           <Route path={routes.ressources_category.path} element={<Layout component={<Category />} />} />
-          <Route path={routes.planner.path} element={<Layout component={<PlannerApp />} />} />
+          <Route path={routes.weather.path} element={<Layout component={<WeatherApp />} />} />
           {/* Ressources routes */}
           {/* <Route path={routes.ressource_details.path} element={<Layout component={<RessourceDetails />} />} /> */}
           <Route path={routes.ressource.path} element={<Layout component={<RessourceCommutator />} />} />
