+++ src/components/Navbar.tsx
@@ -40,7 +40,7 @@ export default function Navbar() {
         <Link className="link" to={routes.home.path}>{routes.home.label}</Link>
         <Link className="link" to={routes.about.path}>{routes.about.label}</Link>
         <Link className="link" to={routes.hub.path}>{routes.hub.label}</Link>
-        <Link className="link" to={routes.planner.path}>{routes.planner.label}</Link>
+        <Link className="link" to={routes.weather.path}>{routes.weather.label}</Link>
         <Link className="link" to={routes.gallery.path}>{routes.gallery.label}</Link>
       </div>
       <div className="navbar__links" style={{ width: '20%', justifyContent: 'flex-end' }}>
@@ -57,7 +57,7 @@ export default function Navbar() {
         <Link className="link" to={routes.home.path} onClick={() => setMobileClick(false)}>{routes.home.label}</Link>
         <Link className="link" to={routes.about.path} onClick={() => setMobileClick(false)}>{routes.about.label}</Link>
         <Link className="link" to={routes.hub.path} onClick={() => setMobileClick(false)}>{routes.hub.label}</Link>
-        <Link className="link" to={routes.planner.path} onClick={() => setMobileClick(false)}>{routes.planner.label}</Link>
+        <Link className="link" to={routes.weather.path} onClick={() => setMobileClick(false)}>{routes.weather.label}</Link>
         <Link className="link" to={routes.gallery.path} onClick={() => setMobileClick(false)}>{routes.gallery.label}</Link>
         <Link className="link" to={routes.contact.path} onClick={() => setMobileClick(false)}>{routes.contact.label}</Link>
         {user && <Link className="link" to={routes.profile.path} onClick={() => setMobileClick(false)}>{routes.profile.label}</Link>}
