+++ src/routes.ts
@@ -23,13 +23,9 @@ export const routes = {
     path: '/ressources/:category/:ressource_slug',
     label: 'Ressource',
   },
-  planner: {
-    path: '/planner',
+  weather: {
+    path: '/weather',
     label: 'Météo en direct',
-  },
-  planner_app: {
-    path: '/planner/app',
-    label: 'Vérifier la météo',
   },
   contact: {
     path: '/contact',
