+++ src/styles/components/app/airItem.scss
@@ -1,34 +1,33 @@
 @import '../../globals.scss';
-@import '../../pages/planner/app.scss';
+@import '../../pages/app/app.scss';
 
-.moon-main-value{
+.air-item{
   display: flex;
   flex-direction: column;
+
+  &.large{
+    align-items: center;
+  }
 
   p{
     margin: 0;
     padding: 0;
   }
 
-  &__title{
-    margin-left: 10px;
-    text-transform: uppercase;
+  .name{
     font-size: .8rem;
 
     &.large{
       font-size: 1rem;
+      margin-bottom: -10px;
     }
   }
 
-  &__value{
+  .value{
     font-size: 1rem;
     font-weight: 800;
 
     &.large{
-      font-size: 1.2rem;
-    }
-
-    &.extra{
       font-size: 2rem;
     }
   }
