+++ src/styles/components/app/moonValue.scss
@@ -1,33 +1,34 @@
 @import '../../globals.scss';
-@import '../../pages/planner/app.scss';
+@import '../../pages/app/app.scss';
 
-.air-item{
+.moon-main-value{
   display: flex;
   flex-direction: column;
-
-  &.large{
-    align-items: center;
-  }
 
   p{
     margin: 0;
     padding: 0;
   }
 
-  .name{
+  &__title{
+    margin-left: 10px;
+    text-transform: uppercase;
     font-size: .8rem;
 
     &.large{
       font-size: 1rem;
-      margin-bottom: -10px;
     }
   }
 
-  .value{
+  &__value{
     font-size: 1rem;
     font-weight: 800;
 
     &.large{
+      font-size: 1.2rem;
+    }
+
+    &.extra{
       font-size: 2rem;
     }
   }
