+++ src/styles/components/app/panels/nextStep.scss
