+++ src/styles/components/app/panels/weather/airQuality.scss
@@ -1,5 +1,5 @@
 @import '../../../../globals.scss';
-@import '../../../../pages/planner/app.scss';
+@import '../../../../pages/app/app.scss';
 
 .air-quality{
   @extend .dashboard-box;
