+++ src/styles/components/app/panels/weather/dailyForecast.scss
@@ -1,5 +1,5 @@
 @import '../../../../globals.scss';
-@import '../../../../pages/planner/app.scss';
+@import '../../../../pages/app/app.scss';
 
 .daily-forecast{
   @extend .dashboard-box;
