+++ src/styles/components/app/panels/weather/hourlyForecast.scss
@@ -1,5 +1,5 @@
 @import '../../../../globals.scss';
-@import '../../../../pages/planner/app.scss';
+@import '../../../../pages/app/app.scss';
 
 .hourly-forecast{
   @extend .dashboard-box;
