+++ src/styles/components/app/panels/weather/moonOverview.scss
@@ -1,5 +1,5 @@
 @import '../../../../globals.scss';
-@import '../../../../pages/planner/app.scss';
+@import '../../../../pages/app/app.scss';
 
 .moon-overview{
   @extend .dashboard-box;
