+++ src/styles/components/app/panels/weather/sunOverview.scss
@@ -1,5 +1,5 @@
 @import '../../../../globals.scss';
-@import '../../../../pages/planner/app.scss';
+@import '../../../../pages/app/app.scss';
 
 .sun-overview{
   @extend .dashboard-box;
