+++ src/styles/components/app/panels/weather/weatherOverview.scss
@@ -1,5 +1,5 @@
 @import '../../../../globals.scss';
-@import '../../../../pages/planner/app.scss';
+@import '../../../../pages/app/app.scss';
 
 .weather-overview{
   @extend .dashboard-box;
