+++ src/styles/pages/app/app.scss
@@ -1,6 +1,6 @@
 @import '../../globals.scss';
 
-.planner-app {
+.weather-app {
   @extend .global-page;
   padding-top: 0;
   padding-bottom: 10vh;
