+++ src/styles/pages/app/weather.scss
